# backend/docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact:
    name: Delicious Team
  description: |-
    家庭菜谱管理系统 — 支持菜谱版本控制、百科搜索、翻译、导入导出
    所有接口前缀 /api/v1，单用户模式（无需认证）
  title: 人间烟火 (Delicious) API
  version: "2.0"
paths:
  /admin/cleanup-images:
    post:
      description: 对比数据库引用和磁盘文件，列出未被引用的孤立图片。仅本地存储模式可用
      produces:
      - application/json
      responses:
        "200":
          description: 'result: total_files, orphan_files, freed_bytes'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Blob 存储不支持此功能
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 扫描孤立图片
      tags:
      - 系统管理
  /admin/cleanup-images/execute:
    post:
      description: 删除所有未被引用的孤立图片。仅本地存储模式可用
      produces:
      - application/json
      responses:
        "200":
          description: 'result: total_files, orphan_files(已删除), freed_bytes'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Blob 存储不支持此功能
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 执行图片清理
      tags:
      - 系统管理
...
  /encyclopedia/{id}/tags:
    get:
      description: 查看某个百科菜谱的所有标签
      parameters:
      - description: 百科菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: items
          schema:
            additionalProperties: true
            type: object
      summary: 获取菜谱关联的标签
      tags:
      - 标签管理
    post:
      consumes:
      - application/json
      description: 为百科菜谱关联一个标签
      parameters:
      - description: 百科菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      - description: '{\'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            format: int64
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: '{}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: tag_id 无效
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 给菜谱添加标签
      tags:
      - 标签管理
  /encyclopedia/{id}/tags/{tag_id}:
    delete:
      description: 取消百科菜谱与标签的关联
      parameters:
      - description: 百科菜谱 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签 ID
        in: path
        name: tag_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{}'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 关联不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 移除菜谱的标签
      tags:
      - 标签管理
...
  /tags:
    get:
      description: 获取全部标签（字典表）
      produces:
      - application/json
      responses:
        "200":
          description: items
          schema:
            additionalProperties: true
            type: object
      summary: 标签列表
      tags:
      - 标签管理
    post:
      consumes:
      - application/json
      description: 新增一个标签名称
      parameters:
      - description: '{\'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: tag
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 名称不能为空
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建标签
      tags:
      - 标签管理
  /tags/{id}:
    delete:
      description: 删除一个标签（同时清除关联关系）
      parameters:
      - description: 标签 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{}'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 标签不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除标签
      tags:
      - 标签管理
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传一张图片（JPG/PNG/WebP/GIF），最大 10MB。Vercel 环境自动使用 Blob 存储
      parameters:
      - description: 图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: url, filename, size
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 文件格式不支持或超过大小限制
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 上传单张图片
      tags:
      - 图片上传
  /upload/batch:
    post:
      consumes:
      - multipart/form-data
      description: 一次上传多张图片，返回全部上传结果
      parameters:
      - collectionFormat: csv
        description: 图片文件数组
        in: formData
        items:
          type: file
        name: files
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: files
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 文件格式不支持
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 批量上传图片
      tags:
      - 图片上传
produces:
- application/json
schemes:
- http
- https
swagger: "2.0"